make router lib file
